Scraper/requirements.txt
tweepy>=4.14.0
praw>=7.7.0
atproto>=0.0.50
sqlalchemy>=2.0
python-dotenv>=1.0
pandas>=2.0
apscheduler>=3.10
